Con Shorts!

+++ 2025/INTRO/Array_Promedio.c
@@ -1,7 +1,7 @@
-/* 
+/*
  Ejercicio básico en lenguaje C donde el usuario carga un vector (array) de 10 posiciones
  y luego se calcula y muestra el promedio de los valores ingresados
- 
+
  Vemos:
  - Uso de arrays
  - Uso de bucles for
@@ -13,14 +13,14 @@
 #include <stdio.h>
 
 int main() {
-    int numeros[10];          // Vector de 10 enteros
-    int suma = 0;             // Variable para acumular la suma
+    short numeros[10];          // Vector de 10 enteros
+    short suma = 0;             // Variable para acumular la suma
     float promedio;           // Variable para almacenar el promedio
 
     // Cargar el vector con valores ingresados por el usuario
     printf("Ingrese 10 numeros enteros:\n");
-    for(int i = 0; i < 10; i++) {
-        printf("Numero %d: ", i + 1);
+    for(short i = 0; i < 10; i++) {
+        printf("Numero %hd: ", i + 1);
         scanf("%d", &numeros[i]);
         suma += numeros[i];  // Sumar el número ingresado
     }
